update dual-sentry heat controller params

+++ src/core/utils/system_ctrl/rmcs_core/src/controller/shooting/shooting_recorder.cpp
@@ -89,7 +89,7 @@ public:
 
 private:
     /// @brief Component interface
-    InputInterface<float> initial_speed_;
+    InputInterface<double> initial_speed_;
     InputInterface<double> shoot_timestamp_;
 
     InputInterface<double> fractional_temperature_;
